Update rot_pseudo2d_anaRB.py
Incorrect multiplication symbol in the calculation of E_def. Should be a '+' instead. Sorry about this!

If i recall correctly it was something I'd fixed for the corresponding viscous dissipation term in the entropy diffusion equation, but I obviously didn't carry this over into the analysis output for E_def and as I never looked at any E_def outputs for the 2.5d code it's something I'd completely missed!